--- .babelrc
@@ -1,10 +1,0 @@
-{
-  "sourceMaps": true,
-  "presets": [
-    ["es2015", { "loose":true }],
-    "stage-0"
-  ],
-  "plugins": [
-    ["transform-react-jsx", { "pragma": "h" }]
-  ]
-}
