+++ package.json
@@ -3,56 +3,17 @@
   "version": "1.0.0",
   "description": "Demo of preact rendering an entire app in a Web Worker.",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --inline --hot --progress",
-    "start": "superstatic build -p 8080 --host 0.0.0.0 --gzip -c '{\"rewrites\": [{\"source\":\"**\",\"destination\":\"index.html\"}],\"headers\":[{\"source\":\"**\",\"headers\":[{\"key\":\"Cache-Control\",\"value\":\"max-age=86400\"}]}]}'",
-    "prestart": "npm run build",
-    "build": "cross-env NODE_ENV=production webpack -p --progress",
-    "prebuild": "mkdirp build && ncp src/assets build/assets",
+    "dev": "npx serve src",
     "test": "eslint {src,test}"
   },
   "keywords": [
     "preact",
     "worker"
   ],
-  "license": "MIT",
-  "author": "Jason Miller <jason@developit.ca>",
   "devDependencies": {
-    "autoprefixer": "^6.4.0",
-    "babel": "^6.5.2",
-    "babel-core": "^6.14.0",
     "babel-eslint": "^7.0.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-react-jsx": "^6.8.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.14.0",
-    "babel-runtime": "^6.11.6",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.26.1",
-    "eslint": "^3.0.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.22.0",
-    "json-loader": "^0.5.4",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "mkdirp": "^0.5.1",
-    "ncp": "^2.0.0",
-    "path": "^0.12.7",
-    "postcss-loader": "^1.1.1",
-    "raw-loader": "^0.5.1",
-    "source-map-loader": "^0.1.5",
-    "style-loader": "^0.13.1",
-    "superstatic": "^4.0.3",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.15.0",
-    "worker-loader": "^0.7.1"
+    "eslint": "^3.0.1"
   },
-  "dependencies": {
-    "ionic-sdk": "^1.2.4",
-    "preact": "^7.1.0",
-    "preact-router": "^2.0.0",
-    "progress-spinner": "^1.1.0"
-  }
+  "license": "MIT",
+  "author": "Jason Miller <jason@developit.ca>"
 }
