+++ src/app.mjs
@@ -1,0 +1,5 @@
+import './renderer/worker.mjs';
+import { h, render } from './deps/preact.mjs';
+import App from './components/app.mjs';
+
+render(h(App), document.body);
