+++ src/deps/htm.mjs
@@ -1,0 +1,2 @@
+export { default } from "https://unpkg.com/htm@3.1.1?module";
+export * from "https://unpkg.com/htm@3.1.1?module";
