+++ src/deps/preact.mjs
@@ -1,0 +1,1 @@
+export * from "https://unpkg.com/preact@7.2.1?module";
