+++ src/index.html
@@ -7,6 +7,8 @@
 		<meta name="mobile-web-app-capable" content="yes">
 		<meta name="apple-mobile-web-app-capable" content="yes">
 		<meta name="format-detection" content="telephone=no">
+		<link rel="stylesheet" href="../style/index.css">
+		<script type="module" src="./index.mjs"></script>
 	</head>
 	<body></body>
 </html>
