--- src/index.js
@@ -1,6 +1,0 @@
-import './style';
-import dom from './renderer/dom';
-import spawnWorker from 'worker-loader!./app';
-
-let worker = spawnWorker();
-dom({ worker });
