+++ src/index.mjs
@@ -1,0 +1,4 @@
+import dom from './renderer/dom.mjs';
+
+const worker = new Worker('./app.mjs', { type: 'module' });
+dom({ worker });
