+++ src/lib/htm.mjs
@@ -1,0 +1,4 @@
+import { h } from "../deps/preact.mjs";
+import htm  from "../deps/htm.mjs";
+
+export const html = htm.bind(h);
