+++ src/style/index.css
@@ -1,0 +1,112 @@
+@import url("https://unpkg.com/progress-spinner@1.1.0/style.css");
+@import url("https://unpkg.com/ionic-sdk@1.2.4/release/css/ionic.css");
+
+html,
+body {
+  height: 100%;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  overflow: hidden;
+}
+
+.app {
+  height: 100%;
+  overflow: hidden;
+}
+
+.app > .content {
+  position: absolute;
+  left: 0;
+  top: 0;
+  bottom: 0;
+  width: 100%;
+  overflow: auto;
+  overflow-x: hidden;
+  overflow-scrolling: touch;
+  -webkit-overflow-scrolling: touch;
+}
+
+.app > .content.has-header {
+  top: 44px;
+}
+
+.app > .content.has-footer {
+  bottom: 44px;
+}
+
+.app > .content .item:not(.item-divider) {
+  position: relative;
+  contain: strict;
+  height: 52px;
+  overflow: hidden;
+}
+
+.unread {
+  animation: 150ms zoomIn forwards 1 ease;
+}
+
+@keyframes zoomIn {
+  from {
+    transform: scale(0.01);
+  }
+}
+
+.bar-header progress-spinner {
+  position: absolute;
+  right: 10px;
+  top: 10px;
+  font-size: 22px;
+}
+
+.dialog {
+  position: fixed;
+  display: none;
+  left: 0;
+  top: 0;
+  width: 0;
+  height: 0;
+  overflow: visible;
+  z-index: 999;
+}
+
+.dialog.open {
+  display: block;
+}
+
+.dialog:before {
+  content: "";
+  position: fixed;
+  display: block;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+}
+
+.dialog > .inner {
+  position: fixed;
+  left: 50%;
+  top: 50%;
+  padding: 20px;
+  width: 300px;
+  max-width: 90%;
+  background: #fff;
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.7);
+  border-radius: 3px;
+  transform: translate(-50%, -50%);
+  animation: 250ms jumpIn forwards 1 ease;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+}
+
+@keyframes jumpIn {
+  from {
+    opacity: 0;
+    transform: translate(-50%, -50%) scale3d(0.1, 0.1, 0.1);
+  }
+}
