--- src/style/index.less
@@ -1,111 +1,0 @@
-@import (css) '~progress-spinner/style.css';
-@import (css) '~ionic-sdk/release/css/ionic.css';
-
-html, body {
-	height: 100%;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-	overflow: hidden;
-}
-
-.app {
-	height: 100%;
-	overflow: hidden;
-
-	> .content {
-		position: absolute;
-		left: 0;
-		top: 0;
-		bottom: 0;
-		width: 100%;
-		overflow: auto;
-		overflow-x: hidden;
-		overflow-scrolling: touch;
-		-webkit-overflow-scrolling: touch;
-
-		&.has-header {
-			top: 44px;
-		}
-		&.has-footer {
-			bottom: 44px;
-		}
-
-		.item:not(.item-divider) {
-			position: relative;
-			contain: strict;
-			height: 52px;
-			overflow: hidden;
-		}
-	}
-}
-
-
-.unread {
-	animation: 150ms zoomIn forwards 1 ease;
-}
-@keyframes zoomIn {
-	from {
-		transform: scale(0.01);
-	}
-}
-
-
-
-.bar-header progress-spinner {
-	position: absolute;
-	right: 10px;
-	top: 10px;
-	font-size: 22px;
-}
-
-
-.dialog {
-	position: fixed;
-	display: none;
-	left: 0;
-	top: 0;
-	width: 0;
-	height: 0;
-	overflow: visible;
-	z-index: 999;
-
-	&.open {
-		display: block;
-	}
-
-	&:before {
-		content: '';
-		position: fixed;
-		display: block;
-		left: 0;
-		top: 0;
-		width: 100%;
-		height: 100%;
-		background: rgba(0,0,0,0.5);
-	}
-
-	> .inner {
-		position: fixed;
-		left: 50%;
-		top: 50%;
-		padding: 20px;
-		width: 300px;
-		max-width: 90%;
-		background: #FFF;
-		box-shadow: 0 5px 15px rgba(0,0,0,0.7);
-		border-radius: 3px;
-		transform: translate(-50%, -50%);
-		animation: 250ms jumpIn forwards 1 ease;
-	}
-}
-
-@keyframes fadeIn {
-	from { opacity:0; }
-}
-
-@keyframes jumpIn {
-	from {
-		opacity: 0;
-		transform: translate(-50%, -50%) scale3d(.1, .1, .1);
-	}
-}
